requirements.txt
accelerate==0.20.3
fastapi==0.95.1
gradio==3.23.0
httpx==0.24.0
markdown-it-py==2.2.0
numpy==1.24.3
prompt-toolkit==3.0.38
pydantic==1.10.7
requests==2.30.0
rich==13.3.5
rouge==1.0.1
sentencepiece==0.1.99
shortuuid==1.0.11
tiktoken==0.4.0
tokenizers==0.13.3
torch>=1.12.0
transformers==4.28.1
uvicorn==0.22.0
bitsandbytes==0.38.1
peft==0.3.0
langchain==0.0.229
deepspeed==0.9.2
sentence_transformers==2.2.2
tensorboard
openai
scipy
termcolor
flask
flask_cors
sentence_transformers
hydra-core
outlines @ git+https://github.com/evgri243/outlines.git@evgri243/relax-model-loading
cbor2
--extra-index-url https://packages.jetbrains.team/pypi/p/grazi/jetbrains-ai-platform-public/simple
grazie-api-gateway-client==0.0.18
